Get a minimal working version with pagination for long histories
During testing at this stage, the git repository was cloned and the server binary
served the application using the git object in the .git folder.

It was noticed that for git's history going back to 2005 with Linus Torvalds' initial
commit till 2018 the latest commit, the browser took too long to load, and in
some cases, failed.

Other adjustments were made - such as using the tabs to show the history items
and renaming the branches menu item to refs - and showing all refs - branches and tags